romboot devel 16

+++ nexell_riscv/tools/bootgen/makebingen.sh
@@ -1,12 +1,23 @@
 #!/bin/bash
 
-python bootbingen.py
+argc=$#
+BINTYPE=$1
+
+if [ $argc -lt 1 ]
+then
+    echo "Invalid argument check usage please"
+    exit
+fi
+  
+python bootbingen.py ${BINTYPE}
 
 #create sdboot.bin 
-dd if=/dev/zero of=sdboot.bin bs=512 count=1
+#dd if=/dev/zero of=sdboot.bin bs=512 count=1
 
 #append nsih.bin to sdboot.bin
 dd if=nsih.bin bs=512 >> sdboot.bin
 
 #append bl1.bin to sdboot.bin
 dd if=bl1.bin >> sdboot.bin
+
+python converthex.py
